aria-quantum-interpreter/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azure-storage-blob
azure-cosmosdb-table
azure-servicebus
azure-identity
azure-keyvault-secrets
numpy
scipy
matplotlib
qiskit
cirq
pennylane
requests
pandas
flask